\name{GeneList_BTA1}
\alias{GeneList_BTA1}
\docType{data}
\title{
Output from GetGeneList function for BTA 1
}
\description{
This output is provided to run the \code{\link{MapMarkers}} function example. It was generated using the \code{\link{GetGeneList}} function example and truncated to only include BTA 1 data.
}
\usage{data(GeneList_BTA1)}
\format{
  The format is:
 chr "GeneList_BTA1"
}
\examples{
data(GeneList_BTA1)
}
\keyword{datasets}
